feat: separated node from lit-node-client
