+++ packages/node-client/README.md
@@ -1,0 +1,19 @@
+# node-client
+
+This library was generated with [Nx](https://nx.dev).
+
+
+
+## Building
+
+Run `nx build node-client` to build the library.
+
+
+
+
+
+## Running unit tests
+
+Run `nx test node-client` to execute the unit tests via [Jest](https://jestjs.io).
+
+
