+++ packages/node-client/src/index.ts
@@ -1,0 +1,5 @@
+export * from './lib/node-client';
+
+// ----- autogen:polyfills:start  -----
+//
+// ----- autogen:polyfills:end  -----
